:hammer: Rework id of player into an int.

+++ server/sources/client/eject.c
@@ -43,7 +43,7 @@ char *eject(client_t *client, server_t *server)
         if (client_id == client->id)
             continue;
         remove_player(tile, client_id);
-        for (int i = 0; client && client_ptr->id != (size_t) client_id; ++i)
+        for (int i = 0; client && client_ptr->id != client_id; ++i)
             client_ptr = get_list(server->clients, i);
         eject_client(client_ptr, server->map, client->position);
     }
